Use latest image for now and move agents to pkg
