+++ pkg/tools/bingsearch/bing.go
