+++ pkg/tools/bingsearch/bing_test.go
