+++ pkg/tools/bingsearch/client.go
