+++ pkg/tools/bingsearch/doc.go
