+++ pkg/tools/weather/internal/client.go
